+++ packages/natives/native/index.js
@@ -231,15 +231,19 @@ export const hasMatch = nativeBindings.hasMatch;
 export const highlightCode = nativeBindings.highlightCode;
 export const htmlToMarkdown = nativeBindings.htmlToMarkdown;
 export const invalidateFsScanCache = nativeBindings.invalidateFsScanCache;
+export const isoBackend = nativeBindings.isoBackend;
+export const isoDiff = nativeBindings.isoDiff;
+export const isoIsUnavailableError = nativeBindings.isoIsUnavailableError;
+export const isoProbe = nativeBindings.isoProbe;
+export const isoResolve = nativeBindings.isoResolve;
+export const isoStart = nativeBindings.isoStart;
+export const isoStop = nativeBindings.isoStop;
 export const listWorkspace = nativeBindings.listWorkspace;
 export const matchesKey = nativeBindings.matchesKey;
 export const matchesKittySequence = nativeBindings.matchesKittySequence;
 export const matchesLegacySequence = nativeBindings.matchesLegacySequence;
 export const parseKey = nativeBindings.parseKey;
 export const parseKittySequence = nativeBindings.parseKittySequence;
-export const projfsOverlayProbe = nativeBindings.projfsOverlayProbe;
-export const projfsOverlayStart = nativeBindings.projfsOverlayStart;
-export const projfsOverlayStop = nativeBindings.projfsOverlayStop;
 export const readImageFromClipboard = nativeBindings.readImageFromClipboard;
 export const sanitizeText = nativeBindings.sanitizeText;
 export const search = nativeBindings.search;
@@ -283,6 +287,21 @@ export const ImageFormat = {
 	JPEG: 1,
 	WEBP: 2,
 	GIF: 3,
+};
+export const IsoBackendKind = {
+	Apfs: 0,
+	Btrfs: 1,
+	Zfs: 2,
+	LinuxReflink: 3,
+	Overlayfs: 4,
+	WindowsBlockClone: 5,
+	Projfs: 6,
+	Rcopy: 7,
+};
+export const IsoChangeKind = {
+	Added: 0,
+	Modified: 1,
+	Removed: 2,
 };
 export const KeyEventType = {
 	Press: 1,